requirements.txt
ultralytics
torchmetrics[detection]